+++ assets/js/blocks/cart-checkout/checkout/form/payment-method-step.js
@@ -12,7 +12,11 @@ import {
 	usePaymentMethods,
 	useStoreCart,
 } from '@woocommerce/base-hooks';
-import { PaymentMethods } from '@woocommerce/base-components/payment-methods';
+
+/**
+ * Internal dependencies
+ */
+import { PaymentMethods } from '../../payment-methods';
 
 const PaymentMethodStep = () => {
 	const { isProcessing: checkoutIsProcessing } = useCheckoutContext();
