+++ assets/js/blocks/cart-checkout/payment-methods/style.scss
